+++ permyt/__init__.py
@@ -1,5 +1,5 @@
 from .api import PermytClient
 
-__version__ = "0.1.5"
+__version__ = "0.1.6"
 
 __all__ = ("PermytClient", "__version__")