+++ permyt/mixins/requests/__init__.py
@@ -1,5 +1,5 @@
 from .connect import UserConnectMixin
-from .disconnect import UserDisconnectMixin
+from .logs import LogsMixin
 from .provider import ProviderMixin
 from .requester import RequesterMixin
 from .scopes import ScopeManagementMixin
@@ -9,7 +9,7 @@ __all__ = (
     "RequesterMixin",
     "ProviderMixin",
     "UserConnectMixin",
-    "UserDisconnectMixin",
     "ScopeManagementMixin",
+    "LogsMixin",
     "InboundMixin",
 )